% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_billType_stockIn_common}
\alias{mo_billType_stockIn_common}
\title{生产订单的单据类型-直接入库普通生产}
\usage{
mo_billType_stockIn_common()
}
\value{
返回值
}
\description{
生产订单的单据类型-直接入库普通生产
}
\examples{
mo_billType_stockIn_common()
}
