% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_billType_stockIn_reWork}
\alias{mo_billType_stockIn_reWork}
\title{生产订单单据类型-直接入库返工生产}
\usage{
mo_billType_stockIn_reWork()
}
\value{
返回单据类型代码
}
\description{
生产订单单据类型-直接入库返工生产
}
\examples{
mo_billType_stockIn_reWork()
}
