% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_billType_stockIn_zz}
\alias{mo_billType_stockIn_zz}
\title{生产订单单据类型-直接入库组装车间专用}
\usage{
mo_billType_stockIn_zz()
}
\value{
返回单据类型代码
}
\description{
生产订单单据类型-直接入库组装车间专用
}
\examples{
mo_billType_stockIn_zz()
}
