% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_release}
\alias{mo_release}
\title{生产订单下达}
\usage{
mo_release(
  config_path = "data-raw/conf.ini",
  kdc_url = "http://47.103.221.12:8126/k3cloud",
  FBillList = list("MO033473")
)
}
\arguments{
\item{config_path}{配置文件}

\item{kdc_url}{网址}

\item{FBillList}{单据列表}
}
\value{
返回值
}
\description{
生产订单下达
}
\examples{
mo_release()
}
