% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_reSave_auto}
\alias{mo_reSave_auto}
\title{自动对新增或重新审核的生产订单再次保存}
\usage{
mo_reSave_auto(
  config_path = "data-raw/conf.ini",
  kdc_url = "http://47.103.221.12:8126/k3cloud",
  FBillType = "SCDD01_SYS"
)
}
\arguments{
\item{config_path}{配置}

\item{kdc_url}{网址}

\item{FBillType}{单据类型}
}
\value{
返回值
}
\description{
自动对新增或重新审核的生产订单再次保存
}
\examples{
mo_reSave_auto()
}
