% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_reSave_batch}
\alias{mo_reSave_batch}
\title{生产订单批量重新保存}
\usage{
mo_reSave_batch(
  config_path = "data-raw/conf.ini",
  kdc_url = "http://47.103.221.12:8126/k3cloud",
  FBillNoList = list("MO032949", "MO033541"),
  FBillType = "SCDD01_SYS"
)
}
\arguments{
\item{config_path}{配置文件}

\item{kdc_url}{网址}

\item{FBillNoList}{单据列表}

\item{FBillType}{单据类型}
}
\value{
返回值
}
\description{
生产订单批量重新保存
}
\examples{
mo_reSave_batch()
}
