% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_update}
\alias{mo_update}
\title{生产订单更新}
\usage{
mo_update(
  config_path = "data-raw/conf.ini",
  kdc_url = "http://47.103.221.12:8126/k3cloud",
  data = mo_testData(),
  FBillType = "SCDD01_SYS"
)
}
\arguments{
\item{config_path}{配置文件}

\item{kdc_url}{网址}

\item{data}{数据}

\item{FBillType}{单据类型}
}
\value{
返回值
}
\description{
生产订单更新
}
\examples{
mo_update()
}
