% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_view}
\alias{mo_view}
\title{生产订单按单查看并格式化结果数据}
\usage{
mo_view(
  config_path = "data-raw/conf.ini",
  kdc_url = "http://47.103.221.12:8126/k3cloud",
  FBillNo = "MO033489"
)
}
\arguments{
\item{config_path}{配置文件}

\item{kdc_url}{k3cloud网址}

\item{FBillNo}{单据编号}
}
\value{
返回结果
}
\description{
生产订单按单查看并格式化结果数据
}
\examples{
mo_view()
}
